finished style parsing - NOT TESTED YET!

+++ PDFAnnotation-itext/src/main/java/pl/edu/agh/tgmg/api/buildingBlocks/DocumentMetaData.java
@@ -1,15 +1,15 @@
 package pl.edu.agh.tgmg.api.buildingBlocks;
 
-import com.itextpdf.text.PageSize;
+import pl.edu.agh.tgmg.itext.generators.styles.elements.BoxValues;
+
 import com.itextpdf.text.Rectangle;
-import pl.edu.agh.tgmg.itext.generators.styles.BoxValues;
 
 public interface DocumentMetaData {
     String getTitle();
 
     String getSubject();
 
-    BoxValues<Integer> getMargins();
+    BoxValues<Float> getMargins();
 
     boolean isCreateDate();
